+++ src/main/java/com/intranet/service/FileService.java
@@ -9,7 +9,7 @@ import com.intranet.model.User;
 public interface FileService {
 	public void save(File file);
 	public void update(File file);
-	public File findById(Integer id);
+	public File findById(Long id);
 	Set<File> findByOwner(User owner);
 	File findByParentAndName(Folder parent, String name);
 	public void remove(File file);
